Exception Handler

+++ src/main/java/lk/lochana/pos/service/Impl/CustomerServiceImpl.java
@@ -28,24 +28,36 @@ public class CustomerServiceImpl implements CustomerService{
 
     @Override
     public void saveCustomer(CustomerDto customerdto) {
+        if(cusRepo.existsById(customerdto.getId())){
+            throw new RuntimeException("Customer already exist");
+        }
         Customer customer = modelMapper.map(customerdto, Customer.class);
         cusRepo.save(customer);
     }
 
     @Override
     public void deleteCustomer(String id) {
+        if(!cusRepo.existsById(id)){
+            throw new RuntimeException("No such customer");
+        }
         Optional<Customer> customer = cusRepo.findById(id);
         cusRepo.delete(customer.get());
     }
 
     @Override
     public void updateCustomer(CustomerDto customerdto) {
+        if(!cusRepo.existsById(customerdto.getId())){
+            throw new RuntimeException("No such customer to update");
+        }
         Customer customer = modelMapper.map(customerdto, Customer.class);
         cusRepo.save(customer);
     }
 
     @Override
     public CustomerDto searchCustomer(String id) {
+        if(!cusRepo.existsById(id)){
+            throw new RuntimeException("No such customer");
+        }
         return modelMapper.map(cusRepo.findById(id).get(), CustomerDto.class);
     }
